contagem dos votos populares

+++ ideax/views.py
@@ -3,6 +3,8 @@ from django.shortcuts import render, get_object_or_404, redirect
 from django.utils import timezone
 from django.http import JsonResponse
 from django.template.loader import render_to_string
+from django.views.decorators.csrf import ensure_csrf_cookie, csrf_protect
+from django.views.decorators.http import require_http_methods
 
 from .models import Idea, Phase, Criterion,Popular_Vote
 from .forms import IdeaForm, PhaseForm, CriterionForm
@@ -10,7 +12,8 @@ from .forms import IdeaForm, PhaseForm, CriterionForm
 def idea_list(request):
     ideas = Idea.objects.order_by('creation_date')
     ideas_liked = get_ideas_liked(request)
-    return render(request, 'ideax/idea_list.html', {'ideas': ideas, 'ideas_liked' : list(ideas_liked), 'link_title': True})
+    ideas_created_by_me = get_ideas_created(request)
+    return render(request, 'ideax/idea_list.html', {'ideas': ideas, 'ideas_liked' : list(ideas_liked), 'link_title': True, 'ideas_created_by_me': ideas_created_by_me,})
 
 """
 def idea_detail(request, pk):
@@ -28,17 +31,16 @@ def idea_detail(request, pk):
 def save_idea(request, form, template_name):
     data = dict()
     if request.method == "POST":
-        form = IdeaForm(request.POST)
         if form.is_valid():
             idea = form.save(commit=False)
-            print (idea.id)
             idea.author = request.user
             idea.creation_date = timezone.now()
             idea.phase= Phase.objects.get(name='Crescendo')
             idea.save()
             data['form_is_valid'] = True
             ideas = Idea.objects.order_by('creation_date')
-            data['html_idea_list'] = render_to_string('ideax/idea_list_loop.html', {'ideas': ideas, 'link_title': True,})
+            ideas_created_by_me = get_ideas_created(request)
+            data['html_idea_list'] = render_to_string('ideax/idea_list_loop.html', {'ideas': ideas, 'link_title': True, 'ideas_created_by_me': ideas_created_by_me,})
         else:
             data['form_is_valid'] = False
 
@@ -54,7 +56,10 @@ def idea_new(request):
     else:
         form = IdeaForm()
 
-    return save_idea(request, form, 'ideax/includes/partial_idea_create.html')
+    if request.is_ajax():
+        return save_idea(request, form, 'ideax/includes/partial_idea_create.html')
+    else:
+        return redirect('idea_list')
 
 @login_required
 def idea_edit(request, pk):
@@ -79,8 +84,18 @@ def idea_publish(request, pk):
 @login_required
 def idea_remove(request, pk):
     idea = get_object_or_404(Idea, pk=pk)
-    idea.delete()
-    return redirect('idea_list')
+    data = dict()
+    if request.method == 'POST':
+        idea.delete()
+        data['form_is_valid'] = True
+        ideas = Idea.objects.order_by('creation_date')
+        ideas_created_by_me = get_ideas_created(request)
+        data['html_idea_list'] = render_to_string('ideax/idea_list_loop.html', {'ideas': ideas, 'link_title': True, 'ideas_created_by_me': ideas_created_by_me,})
+    else:
+        context = {'idea' : idea}
+        data['html_form'] = render_to_string('ideax/includes/partial_idea_remove.html', context, request=request,)
+
+    return JsonResponse(data)
 
 @login_required
 def phase_new(request):
@@ -161,15 +176,23 @@ def like_popular_vote(request, pk):
     vote = Popular_Vote.objects.filter(voter=request.user,idea__pk=pk)
 
     idea_ = Idea.objects.get(pk=pk)
+    like_boolean =  request.path.split("/")[3] == "like"
 
     if vote.count() == 0:
-        like = Popular_Vote(like=True,voter=request.user,voting_date=timezone.now(),idea=idea_)
+        like = Popular_Vote(like=like_boolean,voter=request.user,voting_date=timezone.now(),idea=idea_)
         like.save()
     else:
-        vote.delete()
+        if vote[0].like == like_boolean:
+            vote.delete()
+            like_boolean = None
+        else:
+            vote.update(like=like_boolean)
+
 
     data = dict()
-    data['qtde_votes'] = idea_.count_likes()
+    data['qtde_votes_likes'] = idea_.count_likes()
+    data['qtde_votes_dislikes'] = idea_.count_dislikes()
+    data['class'] = like_boolean
 
     return JsonResponse(data)
 
@@ -179,3 +202,11 @@ def get_ideas_liked(request):
         ideas_liked = Popular_Vote.objects.filter(voter=request.user).values_list('idea_id',flat=True)
 
     return ideas_liked
+
+
+def get_ideas_created(request):
+    ideas_created = []
+    if request.user.is_authenticated:
+        ideas_created = Idea.objects.filter(author=request.user).values_list('id',flat=True)
+
+    return ideas_created
